pyproject.toml
[project]
name = "game-script-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-docx>=1.1.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.6.12",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]

[tool.black]
line-length = 88
target-version = ['py313']
exclude = '''
/(
    \.venv
    | openai_api/rpc
)/
'''
